+++ src/dovi/generator.rs
@@ -127,8 +127,13 @@ impl Generator {
 
         self.config = Some(config);
 
-        if let Some(config) = &self.config {
+        if let Some(config) = self.config.as_mut() {
             println!("Generating metadata: {}...", &config.profile);
+
+            // Correct L1 for sources other than XML
+            if self.xml_path.is_none() {
+                config.fixup_l1();
+            }
 
             config.write_rpus(&self.rpu_out)?;