+++ dolby_vision/CHANGELOG.md
@@ -1,4 +1,6 @@
 ## ??
+- Add `clamp_values` function to `ExtMetadataBlockLevel1`.
+- Add `fixup_l1` function to `GenerateConfig`.
 
 ## 1.6.7